requirements.txt
anyio==3.7.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
fastapi==0.95.2
future==0.18.3
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-bigquery==3.11.0
google-cloud-bigquery-storage==2.20.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
graphql-core==3.2.3
greenlet==2.0.2
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
idna==3.4
numpy==1.24.3
packaging==23.1
proto-plus==1.22.2
protobuf==4.23.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.8
python-dateutil==2.8.2
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy-bigquery==1.6.1
sqlalchemy2-stubs==0.0.2a34
sqlmodel==0.0.8
starlette==0.27.0
strawberry-graphql==0.180.5
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
